Remove minimal admin
